% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edsamplr-package.R
\docType{package}
\name{edsamplr-package}
\alias{edsamplr}
\alias{edsamplr-package}
\title{edsamplr: First Semester Statistics Course Data Simulator}
\description{
A tool for statistics educators to generate simulated data for all the major concepts in a first-semester statistics course. Functions include generating K groups/samples for binary, categorical, and quantitative data, matched pairs quantitative data with specified parameters and/or moments. There is also a function for correlated quantitative data with specified moments, slope, and correlation coefficient. The quantitative variables are produced using Fleishman's power transformation method. You provide the parameters, and we provide the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://keahtandon.github.io/edsamplr/}
}

}
\author{
\strong{Maintainer}: Keah Tandon \email{tandonk@mailbox.sc.edu} (\href{https://orcid.org/0009-0009-2320-3082}{ORCID})

Authors:
\itemize{
  \item Keah Tandon \email{tandonk@mailbox.sc.edu} (\href{https://orcid.org/0009-0009-2320-3082}{ORCID})
  \item Michael A. Seaman \email{mseaman@sc.edu}
}

}
\keyword{internal}
