% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_binary.R
\name{generate_binary}
\alias{generate_binary}
\title{Generate samples of binary data}
\usage{
generate_binary(k = 1, n = 100, p = 0.5, group_names = c(1:0), replication = 1)
}
\description{
Generate samples of binary data
}
