% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_categorical.R
\name{generate_categorical}
\alias{generate_categorical}
\title{Generate samples of categorical data}
\usage{
generate_categorical(
  k = 1,
  n = 100,
  groups = 1,
  group_prop = "equal",
  group_names = "default",
  replication = 1
)
}
\arguments{
\item{k}{A numeric vector for the number of variables. The default value is 1.}

\item{n}{A numeric vector for the sample size. The default value is 100.}

\item{groups}{A numeric vector for the number of groups within each variable. The default value is 1. If k>1, you can provide separate group numbers for each k.}

\item{group_prop}{A vector for the proportion of each group in the population. The default value is "equal". For unequal proportions, provide a concatenated vector the same length as the number of groups. If k>1, you can provide separate sets of proportions for each k.}

\item{group_names}{An optional character vector for naming successes and failures. The default values are 1:groups.}

\item{replication}{A numeric vector for the number of times to replicate the sampling. The default value is 1.}
}
\value{
A list containing a data frame ("sample") with n rows and k columns and a matrix ("proportion_summary") of the input and sample proportions for each group and variable.
}
\description{
generate_categorical() is function to generate k variables of sample categorical data with j groups.
}
\examples{
generate_categorical(k = 2, groups = c(2, 3), group_prop = c(.3, .7, .2, .2, .6))

generate_categorical(k = 2, groups = 2, group_prop = c(.3, .7))



}
