% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_correlated.R
\name{generate_correlated}
\alias{generate_correlated}
\title{Generate samples of correlated quantitative data}
\usage{
generate_correlated(
  n = 100,
  mean = 0,
  var = 1,
  skew = 0,
  kurt = 0,
  slope = 1,
  r = 0.5,
  summary = TRUE,
  decimals = 3,
  replication = 1
)
}
\arguments{
\item{n}{A numeric vector for the sample size. The default value is 100.}

\item{mean}{A numeric vector for the population mean of the first variable. The default value is 0.}

\item{var}{A numeric vector for the population variance of the first variable. The default value is 1.}

\item{skew}{A numeric vector for the population skew of the variables. Separate values of skew can be entered. The default value is 0.}

\item{kurt}{A numeric vector for the population kurtosis of the variables. Separate values of kurtosis can be entered. The default value is 0.}

\item{slope}{A numeric vector for the slope of the regression line. The default value is 0.}

\item{r}{A numeric vector for the correlation between the variables. The default value is 0.5}

\item{summary}{A logical vector for whether the return should include summary statistics. The default value is TRUE.}

\item{decimals}{A numeric vector for the number of decimals to round the sample data to. The default value is 3.}

\item{replication}{A numeric vector for the number of times to replicate the sampling. The default value is 1.}
}
\value{
If summary=TRUE, a list containing the correlation of the sample, the slope of the sample, a data frame ("sample") with n rows and 2 columns, and a matrix ("summary") of the input parameters and descriptive statistics. If summary=FALSE, a data frame with n rows and 2 columns.
}
\description{
generate_correlated() is a function to generate samples of correlated quantitative data with specified first four moments and specified slope.
}
\examples{
# example code


}
