% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distribution.R
\name{generate_distribution}
\alias{generate_distribution}
\title{Generate sample data from named distributions}
\usage{
generate_distribution(
  distribution = c("beta", "cauchy", "chi-squared", "exponential", "f", "gamma",
    "laplace", "logistic", "log normal", "normal", "t", "triangular", "uniform",
    "weibull"),
  n = 1000,
  seed = NULL,
  alpha_shape = NULL,
  beta_shape = NULL,
  location = NULL,
  scale = NULL,
  df = NULL,
  df2 = NULL,
  rate = NULL,
  mean_log = NULL,
  sd_log = NULL,
  mean = NULL,
  sd = NULL,
  theta = NULL,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{distribution}{A character vector to identify which distributional shape to sample. Options include \verb{beta,} \verb{cauchy,} \verb{chi-squared,} \verb{exponential,} \verb{f,} \verb{gamma,} \verb{laplace,} \verb{logistic,} \verb{log normal,} \verb{normal,} \verb{t,} \verb{triangular,} \verb{uniform,} and \code{weibull.}}

\item{n}{A numeric vector for the sample size. The default value is 1000.}

\item{seed}{An optional numeric vector to use in set.seed()}

\item{alpha_shape}{A non-negative numeric vector required for the shape of the beta, gamma, and Weibull distributions.}

\item{beta_shape}{A non-negative numeric vector required for the shape of the beta distribution.}

\item{location}{A numeric vector for the location of the Cauchy, Laplace, and logistic distributions.}

\item{scale}{A numeric vector for the scale of the Cauchy, Laplace, logistic, and Weibull distributions.}

\item{df}{A numeric vector required for the degrees of freedom of the chi-squared, F, and t distributions.}

\item{df2}{A numeric vector required for the degrees of freedom of the F distribution.}

\item{rate}{A numeric vector for the rate of the exponential and gamma distributions.}

\item{mean_log}{A numeric vector for the mean (on the log scale) of the log normal distribution.}

\item{sd_log}{A numeric vector for the standard deviation (on the log scale) of the log normal distribution.}

\item{mean}{A numeric vector for the mean of the normal distribution.}

\item{sd}{A numeric vector for the standard deviation of the normal distribution.}

\item{theta}{A numeric vector required for the theta of the triangular distribution.}

\item{min}{A numeric vector for the minimum value of the triangular and uniform distributions.}

\item{max}{A numeric vector for the maximum value of the triangular and uniform distributions.}
}
\value{
A list of
\itemize{
\item \code{data}, a vector of length n containing the generated values
\item \code{summary}, a descriptive matrix of input and sample values of the specified distribution's parameters
}
}
\description{
\code{generate_distribution()} is a wrapper function for generating sample data from named continuous distributions.

The distributional shapes include beta, Cauchy, chi-squared, exponential, F, gamma, Laplace, logistic, log normal, normal, t, triangular, uniform, and Weibull. See additional details below for distribution parameters and default values.
}
\section{Distributional Shapes & Default Values}{


The \strong{beta} distribution requires the alpha and beta shape parameters. There are no default values for this distribution. See \code{\link[stats:Beta]{stats::rbeta()}} for more information.

The \strong{Cauchy} distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See \code{\link[stats:Cauchy]{stats::rcauchy()}} for more information.

The \strong{chi-squared} distribution requires the df parameter. There is no default value for this distribution. See \code{\link[stats:Chisquare]{stats::rchisq()}} for more information.

The \strong{exponential} distribution uses the rate parameter, which has a default value of 1. See \code{\link[stats:Exponential]{stats::rexp()}} for more information.

The \strong{F} distribution requires the df and df2 parameters. There are no default values for this distribution. See \code{\link[stats:Fdist]{stats::rf()}} for more information.

The \strong{gamma} distribution requires the alpha shape and rate parameter. There is no default value for alpha shape, but the default value for rate is 1. See \code{\link[stats:GammaDist]{stats::rgamma()}} for more information.

The \strong{Laplace} distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See \code{\link[VGAM:laplaceUC]{VGAM::rlaplace()}} for more information.

The \strong{logistic} distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See \code{\link[stats:Logistic]{stats::rlogis()}} for more information.

The \strong{log normal} distribution uses the mean log and sd log parameters. The default values are mean log = 0 and sd log = 1. See \code{\link[stats:Lognormal]{stats::rlnorm()}} for more information.

The \strong{normal} distribution uses the mean and sd parameters. The default values are mean = 0 and sd = 1. See \code{\link[stats:Normal]{stats::rnorm()}} for more information.

The \strong{t} distribution requires the df parameter. There is no default value for this distribution. See \code{\link[stats:TDist]{stats::rt()}} for more information.

The \strong{triangular} distribution requires the theta, min, and max parameters. There is no default value for theta, but the other default values are min = 0 and max = 1. See \code{\link[VGAM:triangleUC]{VGAM::rtriangle()}} for more information.

The \strong{uniform} distribution uses the min and max parameters. The default values are min = 0 and max = 1. See \code{\link[stats:Uniform]{stats::runif()}} for more information.

The \strong{Weibull} distribution requires the alpha shape and scale parameter. There is no default value for alpha shape, but the default value for scale is 1. See \code{\link[stats:Weibull]{stats::rweibull()}} for more information.
}

\examples{
generate_distribution(distribution = "laplace")

generate_distribution(distribution = "gamma", alpha_shape = 3)

generate_distribution(distribution = "beta", alpha_shape = 2, beta_shape = 2)

}
