% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heteroscedastic.R
\name{generate_heteroscedastic}
\alias{generate_heteroscedastic}
\title{Generate samples of heteroscedastic data}
\usage{
generate_heteroscedastic(n = 1000, degree)
}
\arguments{
\item{n}{A numeric vector for the sample size. The default value is 1000.}

\item{degree}{A numeric vector for the degree of heteroscedasticity, in which 1 is homoscedastic and 5 is extremely heteroscedastic.}
}
\value{
A data frame with n rows and 2 columns
}
\description{
\code{generate_heteroscedastic()} is a function to generate samples of data with specified degree of heteroscedasticity
}
\examples{
test <- generate_heteroscedastic(degree = 4)
model <- lm(v2 ~ v1, data = test)
plot(model, 2)

test <- generate_heteroscedastic(n = 20, degree = 2)
model <- lm(v2 ~ v1, data = test)
plot(model, 2)

}
