% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matched.R
\name{generate_matched}
\alias{generate_matched}
\title{Generate samples of quantitative matched pair data with correlation and optional standardized effect}
\usage{
generate_matched(
  n = 100,
  mean = 0,
  var = 1,
  skew = 0,
  kurt = 0,
  effect_size = 0.5,
  r = 0.5,
  use_effect_size = TRUE,
  summary = TRUE,
  decimals = 3,
  replication = 1
)
}
\arguments{
\item{n}{A numeric vector for the sample size. The default value is 100.}

\item{mean}{A numeric vector for the population mean of the first variable. The default value is 0.}

\item{var}{A numeric vector for the population variance of the first variable. The default value is 1.}

\item{skew}{A numeric vector for the population skew of the variables. Separate values of skew can be entered. The default value is 0.}

\item{kurt}{A numeric vector for the population kurtosis of the variables. Separate values of kurtosis can be entered. The default value is 0.}

\item{effect_size}{An optional numeric vector for the standardized effect size between groups of data.}

\item{r}{A numeric vector for the correlation between the variables. The default value is 0.5.}

\item{use_effect_size}{A logical vector to indicate whether to use a standardized effect size in generating data for k when k >= 2. The default value is FALSE.}

\item{summary}{A logical vector for whether the return should include summary statistics. The default value is TRUE.}

\item{decimals}{A numeric vector for the number of decimals to round the sample data to. The default value is 3.}

\item{replication}{A numeric vector for the number of times to replicate the sampling. The default value is 1.

@return If summary = TRUE, a list containing a vector ("correlation") with the correlation between the pairs, a matrix ("summary") of the input parameters and descriptive statistics, and a data frame ("sample") with n rows and 3 columns (values from each set of the pair and the difference of X2 and X1). If summary = FALSE, a data frame with n rows and 3 columns.}
}
\description{
\code{generate_matched()} is a function to generate samples of quantitative matched pair data with specified first four moments, correlation, and optional standardized effect size
}
\examples{
generate_matched()

generate_matched(use_effect_size = FALSE)

generate_matched(skew = c(0.4, 0.25))#'

}
