% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Headrick_method.R
\name{headrick_method}
\alias{headrick_method}
\title{Generate data using Headrick's fifth-order polynomial method}
\usage{
headrick_method(
  k,
  n,
  mean,
  var,
  skew,
  kurtosis,
  corr = diag(1, k),
  gam3 = NaN,
  gam4 = NaN
)
}
\arguments{
\item{k}{A numeric vector indicating the number of variables to sample.}

\item{n}{A numeric vector for the sample size.}

\item{mean}{A numeric vector of length k for the means of each variable
in the distribution. The values can be identical.}

\item{var}{A numeric vector of length k for the variances of each variable
in the distribution. The values can be identical.}

\item{skew}{A numeric vector of length k for the skews of each variable
in the distribution. The values can be identical.}

\item{kurtosis}{A numeric vector of length k for the kurtoses of each variable
in the distribution. The values can be identical.}

\item{corr}{A kxk correlation matrix. The default value is diag(1,k).}

\item{gam3}{An optional numeric value for the 5th standardized cumulant.
The default value is computed internally.}

\item{gam4}{An optional numeric value for the 6th standardized cumulant.
The default value is computed internally.}
}
\value{
A data frame of n rows and k columns containing the generated
values.
}
\description{
\code{headrick_method()} uses specified values of mean, variance, skew, kurtosis,
and the fifth and sixth moments to determine the constants for the expected
values of Y and generate a sample of the specified distribution.
}
\examples{
headrick_method(1, 100, 0, 1, 0, 0)

headrick_method(2, 100, c(0,0), c(1,1), c(0,0), c(0,0))

corr <- matrix(c(1, .7, .7, 1), byrow = TRUE, nrow = 2)
headrick_method(2, 100, c(0,0), c(1,1), c(0,0), c(0,0), corr = corr)

}
\references{
Headrick, T. C. (2002). Fast fifth-order polynomial transforms
for generating univariate and multivariate nonnormal distributions.
\emph{Computational Statistics & Data Analysis, 40}, 685-711.

Olvera Astivia, O. L., & Zumbo, B. D. (2015). A Cautionary Note on the Use
of the Vale and Maurelli Method to Generate Multivariate, Nonnormal Data for
Simulation Purposes. \emph{Educational and Psychological Measurement, 75}, 541-567.
}
\author{
Oscar L. Olvera Astivia and Keah Tandon
}
