% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_binary.R
\name{generate_binary}
\alias{generate_binary}
\title{Generate samples of binary data}
\usage{
generate_binary(k = 1, n = 100, p = 0.5, group_names = c(1:0), replication = 1)
}
\arguments{
\item{k}{A numeric vector for the number of variables. The default value is 1.}

\item{n}{A numeric vector for the sample size. The default value is 100.}

\item{p}{A numeric vector for the proportion of successes. The default value is 0.5. If k>1, you can provide separate proportion values for each k.}

\item{group_names}{An optional character vector for naming successes and failures. The default values are 1 for success and 0 for failure.}

\item{replication}{A numeric vector for the number of times to replicate the sampling. The default value is 1.}
}
\value{
A list containing a data frame ("sample") with n rows and k columns, a matrix ("summary") of the sample size, number of successes, proportion statistic, and proportion parameter for each variable, and a matrix ("freq.table") of the frequencies of successes and failures for each variable.
}
\description{
generate_binary() is function to generate sample binary data.
}
\examples{
generate_binary()

generate_binary(k = 2, p = c(0.5, 0.7))

generate_binary(group_names = c("success", "fail"))#'

}
