% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_binary.R
\name{generate_binary}
\alias{generate_binary}
\title{Generate samples of binary data}
\usage{
generate_binary(
  p = 1,
  n = 100,
  prop = 0.5,
  group_names = c(1:0),
  summary = TRUE,
  replication = 1
)
}
\arguments{
\item{p}{A numeric vector for the number of variables. The default value is 1.}

\item{n}{A numeric vector for the sample size. The default value is 100.}

\item{prop}{A numeric vector for the proportion of successes. The default value is 0.5. If k > 1, you can provide separate proportion values for each k.}

\item{group_names}{An optional character vector for naming successes and failures. The default values are 1 for success and 0 for failure.}

\item{summary}{A logical vector for whether the return should include summary statistics. The default value is TRUE.}

\item{replication}{A numeric vector for the number of times to replicate the sampling. The default value is 1.}
}
\value{
If summary = TRUE, a list containing
\itemize{
\item \code{sample}, a data frame with n rows and k columns
\item \code{summary}, a matrix of the sample size, number of successes, proportion statistic, and proportion parameter for each variable, and
\item \code{freq.table}, a matrix of the frequencies of successes and failures for each variable
}

If summary = FALSE, a data frame with n rows and 1 column.
}
\description{
\code{generate_binary()} is function to generate sample binary data.
}
\details{
This function generates samples of binary data based on specified proportions. It can generate a single column or multiple columns of data with different proportions specified for each. The data is sampled from the specified proportions as a parameter, so the output statistics will not be an exact match to the parameters.

Because the output is a sample, the summary argument allows for summary statistics to be generated with the sample data. This allows the user to easily compare the simulation statistics to the specified parameters.

This is useful for introducing students to the basics of probability, proportion tests, and chi square tests without having to call two (or more) separate \code{rbinom()} functions. If you are interested in using binary data for sample student test responses, consider using \code{generate_2PL()}, which simulates binary response data based on specified difficulty and discrimination parameters.
}
\examples{
generate_binary()

generate_binary(p = 2, prop = c(0.5, 0.7))

generate_binary(group_names = c("success", "fail"))

}
