% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distribution.R
\name{generate_distribution}
\alias{generate_distribution}
\title{Generate sample data from named distributions}
\usage{
generate_distribution(
  distribution = c("beta", "cauchy", "chi-squared", "exponential", "f", "gamma",
    "laplace", "logistic", "log normal", "normal", "t", "triangular", "uniform",
    "weibull"),
  n = 100,
  seed = NULL,
  alpha_shape = NULL,
  beta_shape = NULL,
  location = NULL,
  scale = NULL,
  df = NULL,
  df2 = NULL,
  rate = NULL,
  mean_log = NULL,
  sd_log = NULL,
  mean = NULL,
  sd = NULL,
  theta = NULL,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{distribution}{A character vector to identify which distributional shape to sample. Options include \verb{beta,} \verb{cauchy,} \verb{chi-squared,} \verb{exponential,} \verb{f,} \verb{gamma,} \verb{laplace,} \verb{logistic,} \verb{log normal,} \verb{normal,} \verb{t,} \verb{triangular,} \verb{uniform,} and \code{weibull.}}

\item{n}{A numeric vector for the sample size. The default value is 100.}

\item{seed}{An optional numeric vector to use in set.seed()}

\item{alpha_shape}{A non-negative numeric vector required for the shape of the beta, gamma, and Weibull distributions.}

\item{beta_shape}{A non-negative numeric vector required for the shape of the beta distribution.}

\item{location}{A numeric vector for the location of the Cauchy, Laplace, and logistic distributions. The default value is 0.}

\item{scale}{A numeric vector for the scale of the Cauchy, Laplace, logistic, and Weibull distributions. The default value is 1.}

\item{df}{A numeric vector required for the degrees of freedom of the chi-squared, F, and t distributions.}

\item{df2}{A numeric vector required for the degrees of freedom of the F distribution.}

\item{rate}{A numeric vector for the rate of the exponential and gamma distributions. The default value is 1.}

\item{mean_log}{A numeric vector for the mean (on the log scale) of the log normal distribution. The default value is 0.}

\item{sd_log}{A numeric vector for the standard deviation (on the log scale) of the log normal distribution. The default value is 1.}

\item{mean}{A numeric vector for the mean of the normal distribution. The default value is 0.}

\item{sd}{A numeric vector for the standard deviation of the normal distribution. The default value is 1.}

\item{theta}{A numeric vector required for the theta of the triangular distribution.}

\item{min}{A numeric vector for the minimum value of the triangular and uniform distributions. The default value is 0.}

\item{max}{A numeric vector for the maximum value of the triangular and uniform distributions. The default value is 1.}
}
\value{
A list of \verb{data:} a vector of length n containing the generated values, and \verb{summary:} a descriptive matrix of input and sample values of the specified distribution's parameters.
}
\description{
generate_distribution() is a wrapper function for generating sample data from named continuous distributions.

The distributional shapes include beta, Cauchy, chi-squared, exponential, F, gamma, Laplace, logistic, log normal, normal, t, triangular, uniform, and Weibull.
}
\examples{
generate_distribution(distribution = "beta", alpha_shape = 2, beta_shape = 2)

generate_distribution(distribution = "cauchy")

generate_distribution(distribution = "chi-squared", df = 3)

generate_distribution(distribution = "exponential")

generate_distribution(distribution = "f", df = 4, df2 = 30)

generate_distribution(distribution = "gamma", alpha_shape = 3)

generate_distribution(distribution = "laplace")

generate_distribution(distribution = "logistic")

generate_distribution(distribution = "log normal")

generate_distribution(distribution = "normal")

generate_distribution(distribution = "t", df = 2)

generate_distribution(distribution = "triangular", theta = 1, min = 0, max = 2)

generate_distribution(distribution = "uniform")

generate_distribution(distribution = "weibull", alpha_shape = 1)#'

}
